microservices/api_gateway/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.5
